CrazySpawner: fixed Drop getChance return type

+++ CrazySpawner/src/de/st_ddt/crazyutil/Drop.java
@@ -44,7 +44,7 @@ public class Drop implements ConfigurationSaveable
 		return item.clone();
 	}
 
-	public double getChance()
+	public float getChance()
 	{
 		return chance;
 	}
